# perf: Speedup writing of Parquet primitive values