--- a/crates/polars-compute/src/min_max/simd.rs
+++ b/crates/polars-compute/src/min_max/simd.rs
@@ -72,6 +72,87 @@ where
     Some(state)
 }
 
+fn fold_agg_min_max_kernel<const N: usize, T, F>(
+    arr: &[T],
+    validity: Option<&Bitmap>,
+    min_scalar_identity: T,
+    max_scalar_identity: T,
+    mut simd_f: F,
+) -> Option<(Simd<T, N>, Simd<T, N>)>
+where
+    T: SimdElement + NativeType,
+    F: FnMut((Simd<T, N>, Simd<T, N>), (Simd<T, N>, Simd<T, N>)) -> (Simd<T, N>, Simd<T, N>),
+    LaneCount<N>: SupportedLaneCount,
+{
+    if arr.is_empty() {
+        return None;
+    }
+
+    let mut arr_chunks = arr.chunks_exact(N);
+
+    let min_identity = Simd::splat(min_scalar_identity);
+    let max_identity = Simd::splat(max_scalar_identity);
+    let mut state = (min_identity, max_identity);
+    if let Some(valid) = validity {
+        if valid.unset_bits() == arr.len() {
+            return None;
+        }
+
+        let mask = BitMask::from_bitmap(valid);
+        let mut offset = 0;
+        for c in arr_chunks.by_ref() {
+            let m: Mask<_, N> = mask.get_simd(offset);
+            let slice = Simd::from_slice(c);
+            state = simd_f(
+                state,
+                (m.select(slice, min_identity), m.select(slice, max_identity)),
+            );
+            offset += N;
+        }
+        if arr.len() % N > 0 {
+            let mut min_rest: [T; N] = min_identity.to_array();
+            let mut max_rest: [T; N] = max_identity.to_array();
+
+            let arr_rest = arr_chunks.remainder();
+            min_rest[..arr_rest.len()].copy_from_slice(arr_rest);
+            max_rest[..arr_rest.len()].copy_from_slice(arr_rest);
+
+            let m: Mask<_, N> = mask.get_simd(offset);
+
+            let min_rest = Simd::from_array(min_rest);
+            let max_rest = Simd::from_array(max_rest);
+
+            state = simd_f(
+                state,
+                (
+                    m.select(min_rest, min_identity),
+                    m.select(max_rest, max_identity),
+                ),
+            );
+        }
+    } else {
+        for c in arr_chunks.by_ref() {
+            let slice = Simd::from_slice(c);
+            state = simd_f(state, (slice, slice));
+        }
+        if arr.len() % N > 0 {
+            let mut min_rest: [T; N] = min_identity.to_array();
+            let mut max_rest: [T; N] = max_identity.to_array();
+
+            let arr_rest = arr_chunks.remainder();
+            min_rest[..arr_rest.len()].copy_from_slice(arr_rest);
+            max_rest[..arr_rest.len()].copy_from_slice(arr_rest);
+
+            let min_rest = Simd::from_array(min_rest);
+            let max_rest = Simd::from_array(max_rest);
+
+            state = simd_f(state, (min_rest, max_rest));
+        }
+    }
+
+    Some(state)
+}
+
 macro_rules! impl_min_max_kernel_int {
     ($T:ty, $N:literal) => {
         impl MinMaxKernel for PrimitiveArray<$T> {
@@ -91,12 +172,27 @@ macro_rules! impl_min_max_kernel_int {
                 .map(|s| s.reduce_max())
             }
 
+            fn min_max_ignore_nan_kernel(&self) -> Option<(Self::Scalar<'_>, Self::Scalar<'_>)> {
+                fold_agg_min_max_kernel::<$N, $T, _>(
+                    self.values(),
+                    self.validity(),
+                    <$T>::MAX,
+                    <$T>::MIN,
+                    |(cmin, cmax), (min, max)| (cmin.simd_min(min), cmax.simd_max(max)),
+                )
+                .map(|(min, max)| (min.reduce_min(), max.reduce_max()))
+            }
+
             fn min_propagate_nan_kernel(&self) -> Option<Self::Scalar<'_>> {
                 self.min_ignore_nan_kernel()
             }
 
             fn max_propagate_nan_kernel(&self) -> Option<Self::Scalar<'_>> {
                 self.max_ignore_nan_kernel()
+            }
+
+            fn min_max_propagate_nan_kernel(&self) -> Option<(Self::Scalar<'_>, Self::Scalar<'_>)> {
+                self.min_max_ignore_nan_kernel()
             }
         }
 
@@ -113,12 +209,27 @@ macro_rules! impl_min_max_kernel_int {
                     .map(|s| s.reduce_max())
             }
 
+            fn min_max_ignore_nan_kernel(&self) -> Option<(Self::Scalar<'_>, Self::Scalar<'_>)> {
+                fold_agg_min_max_kernel::<$N, $T, _>(
+                    self,
+                    None,
+                    <$T>::MAX,
+                    <$T>::MIN,
+                    |(cmin, cmax), (min, max)| (cmin.simd_min(min), cmax.simd_max(max)),
+                )
+                .map(|(min, max)| (min.reduce_min(), max.reduce_max()))
+            }
+
             fn min_propagate_nan_kernel(&self) -> Option<Self::Scalar<'_>> {
                 self.min_ignore_nan_kernel()
             }
 
             fn max_propagate_nan_kernel(&self) -> Option<Self::Scalar<'_>> {
                 self.max_ignore_nan_kernel()
+            }
+
+            fn min_max_propagate_nan_kernel(&self) -> Option<(Self::Scalar<'_>, Self::Scalar<'_>)> {
+                self.min_max_ignore_nan_kernel()
             }
         }
     };
@@ -152,6 +263,17 @@ macro_rules! impl_min_max_kernel_float {
                 .map(|s| s.reduce_max())
             }
 
+            fn min_max_ignore_nan_kernel(&self) -> Option<(Self::Scalar<'_>, Self::Scalar<'_>)> {
+                fold_agg_min_max_kernel::<$N, $T, _>(
+                    self.values(),
+                    self.validity(),
+                    <$T>::NAN,
+                    <$T>::NAN,
+                    |(cmin, cmax), (min, max)| (cmin.simd_min(min), cmax.simd_max(max)),
+                )
+                .map(|(min, max)| (min.reduce_min(), max.reduce_max()))
+            }
+
             fn min_propagate_nan_kernel(&self) -> Option<Self::Scalar<'_>> {
                 fold_agg_kernel::<$N, $T, _>(
                     self.values(),
@@ -171,6 +293,27 @@ macro_rules! impl_min_max_kernel_float {
                 )
                 .map(|s| scalar_reduce_max_propagate_nan(s.as_array()))
             }
+
+            fn min_max_propagate_nan_kernel(&self) -> Option<(Self::Scalar<'_>, Self::Scalar<'_>)> {
+                fold_agg_min_max_kernel::<$N, $T, _>(
+                    self.values(),
+                    self.validity(),
+                    <$T>::INFINITY,
+                    <$T>::NEG_INFINITY,
+                    |(cmin, cmax), (min, max)| {
+                        (
+                            (cmin.simd_lt(min) | cmin.simd_ne(cmin)).select(cmin, min),
+                            (cmax.simd_gt(max) | cmax.simd_ne(cmax)).select(cmax, max),
+                        )
+                    },
+                )
+                .map(|(min, max)| {
+                    (
+                        scalar_reduce_min_propagate_nan(min.as_array()),
+                        scalar_reduce_max_propagate_nan(max.as_array()),
+                    )
+                })
+            }
         }
 
         impl MinMaxKernel for [$T] {
@@ -186,6 +329,17 @@ macro_rules! impl_min_max_kernel_float {
                     .map(|s| s.reduce_max())
             }
 
+            fn min_max_ignore_nan_kernel(&self) -> Option<(Self::Scalar<'_>, Self::Scalar<'_>)> {
+                fold_agg_min_max_kernel::<$N, $T, _>(
+                    self,
+                    None,
+                    <$T>::NAN,
+                    <$T>::NAN,
+                    |(cmin, cmax), (min, max)| (cmin.simd_min(min), cmax.simd_max(max)),
+                )
+                .map(|(min, max)| (min.reduce_min(), max.reduce_max()))
+            }
+
             fn min_propagate_nan_kernel(&self) -> Option<Self::Scalar<'_>> {
                 fold_agg_kernel::<$N, $T, _>(self, None, <$T>::INFINITY, |a, b| {
                     (a.simd_lt(b) | a.simd_ne(a)).select(a, b)
@@ -198,6 +352,27 @@ macro_rules! impl_min_max_kernel_float {
                     (a.simd_gt(b) | a.simd_ne(a)).select(a, b)
                 })
                 .map(|s| scalar_reduce_max_propagate_nan(s.as_array()))
+            }
+
+            fn min_max_propagate_nan_kernel(&self) -> Option<(Self::Scalar<'_>, Self::Scalar<'_>)> {
+                fold_agg_min_max_kernel::<$N, $T, _>(
+                    self,
+                    None,
+                    <$T>::INFINITY,
+                    <$T>::NEG_INFINITY,
+                    |(cmin, cmax), (min, max)| {
+                        (
+                            (cmin.simd_lt(min) | cmin.simd_ne(cmin)).select(cmin, min),
+                            (cmax.simd_gt(max) | cmax.simd_ne(cmax)).select(cmax, max),
+                        )
+                    },
+                )
+                .map(|(min, max)| {
+                    (
+                        scalar_reduce_min_propagate_nan(min.as_array()),
+                        scalar_reduce_max_propagate_nan(max.as_array()),
+                    )
+                })
             }
         }
     };